Add files via upload
My inaugural project as an intern at InternSavy during the dynamic period of July to August 2023. The project entails the development of a user-friendly Contact Form, meticulously crafted to facilitate the seamless submission of user information and messages. Leveraging the power of HTML and CSS, I've crafted an interactive and aesthetically pleasing contact form that ensures a delightful user experience